fix 0.8

+++ modules/task_1/andronov_m_min_column_matrix/min_column_matrix.cpp
@@ -53,7 +53,7 @@ std::vector <int> GetParallelMinValueColumn(std::vector < std::vector<int> > Mat
 
     int Error = 0;
     if (rank == 0) {
-        if (rows != Matrix.size() || columns != Matrix[0].size()) {
+        if (rows != static_cast<int>(Matrix.size()) || columns != static_cast<int>(Matrix[0].size())) {
             Error = 1;
             for (int i = 1; i < size; i++)
                 MPI_Send(&Error, 1, MPI_INT, i, 0, MPI_COMM_WORLD);
